Fix errors in main.py and get basic AI to work
The AI "works" now. It can play against itself and uses a very basic
evaluation function. Though there seems to be some issues (the AI
doesn't seem to always take the move with the highest evaluation).

Furthermore, it would be extremely useful to get a log of the currently
played game and export it to a more user-friendly visualization like
chess.com.