Add basic benchmarks
Possible improvements:

- benchmark early- mid- lategame
- benchmark `board.push`, i.e. how long it takes to make moves on the
board
- write benchmarks to .csv file

Also visualizing the results with a plot would be great. Here it would
be good to see how the time changes based on how many legal moves there
are. As the project continues (with aplha beta, move ordering, etc.),
the graph of this should flatten out more and more.