restructure(engine): Use integers for board
Using ints instead of strs for the board representation. This is mainly
because it's easier to work with ints. For performance it didn't seem to
really matter which type is used (which also makes sense, given that the
main bottleneck comes from running many iterations).

Also tried to use numpy arrays here. As expected, they were again slower
then simple python lists.