AI: Fix checkmate and stalemate evaluation
There was a foundamental problem with the evaluation: We didn't actually
check if a single board node is checkmate or stalemate (in which case it
is a leaf node).

Because of this, the tests for the play NEVER worked, only for depth 1,
in which case EVERY node was a leaf node.

Now the move checks if a node is a checkmate or stalemate and simply
returns the evaluated board state.